update, add gitignore, modify workflow
